fixing some possible issues related to findbugs

+++ src/main/java/com/flazr/io/f4v/MovieInfo.java
@@ -49,7 +49,7 @@ public class MovieInfo {
     }
 
     public double getDuration() {
-        return mvhd.getDuration() / mvhd.getTimeScale();
+        return mvhd.getDuration() / (double) mvhd.getTimeScale();
     }
 
     private void initSamples() {
